Bump postgres:10 to postgres:12; re-enable gvm-postgres build
Explanation: The real reason for this change is that as of `postgres:12` (the
public Docker image), the base image has been changed from Debian
`stretch-slim` to Debian `buster-slim`. This has solved a dependency issue that
was encountered while trying to install `libgvm-pg-server0`.

Additionally, I haven't yet seen any list of favoured PostgreSQL versions from
Greenbone, so I'm willing to give this a try.